Update TokenManager.java
Token Availability: Ensure that players are assigned tokens only if they are available. Once a player chooses a token, it should be removed from the available tokens list to prevent duplication.

Token Removal on Player Elimination: If a player is eliminated from the game (e.g., bankrupt), their token should be returned to the list of available tokens.

Token Assignment: Tokens should be assigned to players, and any logic for managing the number of available tokens should be handled properly.